feat: update cancel appointments query
