+++ components/practitioners/AppointmentList.tsx
@@ -1,49 +1,12 @@
-import { IoMdCloseCircleOutline } from "react-icons/io";
-import ConfirmAction from "../custom/ConfirmAction";
-import { Button } from "../ui/Button";
 import { dateFormat, timeFormat } from "@/utils/helpers";
 import Image from "next/image";
-import useDisclosure from "@/hooks/useDisclosure";
-import { useCancelAppointment } from "@/api/appointments";
-import config from "@/utils/config";
-import { usePathname, useRouter } from "next/navigation";
-import { useAuthContext } from "@/providers/AuthProvider";
-import { useToast } from "../ui/use-toast";
+import CancelAppointment from "./CancelAppointments";
 
 interface AppointmentItemProps {
     appointment: PractitionerAppointments;
 }
 
 const AppointmentList = ({ appointment }: AppointmentItemProps) => {
-    const router = useRouter();
-    const { isLoggedIn, user } = useAuthContext();
-    const { toast } = useToast();
-    const pathname = usePathname();
-    const { isOpen, onClose, onOpen } = useDisclosure();
-    const {
-        isPending,
-        mutate: Cancel,
-        data: CancelledData 
-    } = useCancelAppointment();
-
-    console.log(CancelledData)
-    const onCancel = (id: number ) => {
-        if (!isLoggedIn || !user?.id) {
-            toast({ description: "Please sign in to continue", variant: "error" });
-            localStorage.setItem(config.key.redirect, pathname);
-            router.push("/auth/sign-in");
-            return;
-        }
-        const payload: PractitionerAppointmentPayload = {
-            status: "cancelled"
-        };
-        Cancel({
-            payload,
-            appointmentId: id.toString(),
-        });
-        onClose();
-    };
-
 return (
     <div
         className="flex items-center flex-col sm:flex-row gap-8 border-b border-gray-200 pb-5"
@@ -95,28 +58,7 @@ return (
             >
                 {appointment.status}
             </p>
-
-            <div className="flex gap-2 items-center">
-                {appointment.status === "pending" && (
-                    <Button
-                        variant="error"
-                        size="sm"
-                        className="text-red-500 text-sm w-full md:w-fit "
-                        onClick={onOpen}
-                    >
-                        <IoMdCloseCircleOutline className="text-base -mr-1" />
-                        Cancel
-                    </Button>
-                )}
-                <ConfirmAction
-                    isOpen={isOpen}
-                    onContinue={() => onCancel(appointment.id)}
-                    onClose={onClose}
-                    onCancel={onClose}
-                    isLoading={isPending}
-                    message={`Are you sure you want to cancel your appointment with ${appointment.user.last_name} ${appointment.user.first_name}?`}
-                />
-            </div>
+            <CancelAppointment appointment={appointment}/>
         </div>
     </div>
 )
